update all text channel generating functions plus almost finshed with focus

+++ commands/spotify.js
@@ -3,14 +3,16 @@ const guld_mngr = require('./../functions/guild_manager');
 module.exports = async (client, message, args, portal_guilds, portal_managed_guilds_path) => {
     if (portal_guilds[message.guild.id].spotify === message.channel.id && args.length === 0) {
         return {
-            result: true, value: '**This is already the Spotify channel.**'
+            result: true, value: '**This already is, the Spotify channel.**'
         };
     } else if (portal_guilds[message.guild.id].announcement !== message.channel.id) {
         if (spotify = message.guild.channels.cache.find(channel =>
             channel.id == portal_guilds[message.guild.id].spotify)) {
-            spotify
-                .delete()
-                .catch(console.error);
+            if (spotify.deletable) {
+                spotify
+                    .delete()
+                    .catch(console.error); // edo petaei error investigate
+            }
         } else {
             portal_guilds[message.guild.id].spotify = null;
         }
@@ -29,15 +31,13 @@ module.exports = async (client, message, args, portal_guilds, portal_managed_gui
         console.log('2spotify_category: ', spotify_category);
         
         if (spotify_channel !== '') {
-console.log('exei category');
             guld_mngr.create_spotify_channel(
                 message.guild, spotify_channel, spotify_category, portal_guilds[message.guild.id]);
 
             return {
                 result: true, value: '*Spotify channel and category have been created*'
             };
-        } else if (spotify_channel === '') {
-console.log('den exei category');
+        } else if (spotify_channel === '' && spotify_category !== '') {
             guld_mngr.create_spotify_channel(
                 message.guild, spotify_category, null, portal_guilds[message.guild.id]);
